+++ src/api/services/ShlinkApiClientBuilder.ts
@@ -23,7 +23,7 @@ export const buildShlinkApiClient = (httpClient: HttpClient) => (getStateOrSelec
     : getStateOrSelectedServer;
   const serverKey = `${apiKey}_${baseUrl}`;
 
-  const apiClient = apiClients[serverKey] ?? new ShlinkApiClient(httpClient, { apiKey, baseUrl });
+  const apiClient = apiClients[serverKey] ?? new ShlinkApiClient(httpClient, { apiKey, baseUrl }, { credentials: 'include' });
   apiClients[serverKey] = apiClient;
 
   return apiClient;