+++ src/servers/services/ServersImporter.ts
@@ -1,14 +1,20 @@
 import type { CsvToJson } from '../../utils/helpers/csvjson';
 import type { ServerData } from '../data';
+import { deserializeServer } from '../data';
 
-const validateServer = (server: any): server is ServerData =>
-  typeof server.url === 'string' && typeof server.apiKey === 'string' && typeof server.name === 'string';
-
-const validateServers = (servers: any): servers is ServerData[] =>
-  Array.isArray(servers) && servers.every(validateServer);
+const validateAndDeserializeServers = (servers: unknown): ServerData[] => {
+  if (!Array.isArray(servers)) {
+    throw new Error('Provided file does not have the right format.');
+  }
+  return servers.map(deserializeServer);
+};
 
 export class ServersImporter {
-  public constructor(private readonly csvToJson: CsvToJson) {}
+  readonly #csvToJson: CsvToJson;
+
+  public constructor(csvToJson: CsvToJson) {
+    this.#csvToJson = csvToJson;
+  }
 
   public async importServersFromFile(file: File | null | undefined): Promise<ServerData[]> {
     if (!file) {
@@ -16,12 +22,8 @@ export class ServersImporter {
     }
 
     const content = await file.text();
-    const servers = await this.csvToJson(content);
+    const servers = await this.#csvToJson(content);
 
-    if (!validateServers(servers)) {
-      throw new Error('Provided file does not have the right format.');
-    }
-
-    return servers;
+    return validateAndDeserializeServers(servers);
   }
 }