+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Update to `@shlinkio/shlink-frontend-kit` 0.7.0
 * Update to `@shlinkio/shlink-web-component` 0.13.0
 * Update to `@shlinkio/shlink-js-sdk` 2.0.0
+* Add `eslint-plugin-react-compiler`
 
 ### Deprecated
 * *Nothing*